fix: remove unsupported --web-addr flag from ngrok command
The --web-addr flag is not supported as a command-line argument in ngrok v3.
It should only be used in the configuration file. This was causing the container
to fail with 'ERROR: unknown flag: --web-addr' and enter a restart loop.

Changed from config file approach to direct http command for simplicity.